aplz_api/requirements.txt
Django==5.0.1
django-cors-headers==4.3.1
django-paranoid-model==1.0.7
djangorestframework==3.15.1
mysqlclient==2.2.4
djangorestframework_simplejwt==5.3.1
requests==2.25.1
django-storages==1.14.2
django-mysql==4.12.0
django-query-capture==0.3.2